requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
pydantic-settings>=2.0.0
yfinance>=0.2.40
pandas>=2.0.0
pandas-ta>=0.3.14b0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6